Add files via upload
This terminal based project includes the File Handling system where it generates the otp and sends that otp on users email. The Main class contains the log in and sign up methods from which user can log in and sign up for every action an otp is generated and sent on users email those who log in or sign in and all the data will be save in the database and the main content will be show after the log in or sign up, such as show the hidden files, hide files, and unhide files where we provide the file path and that file will be stored as the binary format in the database.